requirements.txt.txt
python-telegram-bot
feedparser 